+++ packages/bits/spec/components/form-field/form-field.e2e.ts
@@ -116,8 +116,8 @@ describe("USERCONTROL form-field", () => {
         expect(await select.isSelectDisabled()).toBeTruthy();
         expect(await switchElement.disabled()).toBeTruthy();
         expect(await timepicker.textbox.disabled()).toBeTruthy();
-        await timepicker.popup.getPopupToggle().click();
-        expect(await timepicker.popup.isOpened()).toBeFalsy();
+        await timepicker.toggle();
+        expect(await timepicker.overlay.isOpened()).toBeFalsy();
         expect(await checkbox.isDisabled()).toBeTruthy();
         expect(await dateTimepicker.isDisabled()).toBeTruthy();
     });
@@ -150,8 +150,8 @@ describe("USERCONTROL form-field", () => {
         });
         it("should dynamically enable timepicker", async () => {
             expect(await timepicker.textbox.disabled()).toBeFalsy();
-            await timepicker.popup.getPopupToggle().click();
-            expect(await timepicker.popup.isOpened()).toBeTruthy();
+            await timepicker.toggle();
+            expect(await timepicker.overlay.isOpened()).toBeTruthy();
         });
 
         it("should dynamically enable checkbox", async () => {
@@ -186,8 +186,8 @@ describe("USERCONTROL form-field", () => {
             expect(await select.isSelectDisabled()).toBeTruthy();
             expect(await switchElement.disabled()).toBeTruthy();
             expect(await timepicker.textbox.disabled()).toBeTruthy();
-            await timepicker.popup.getPopupToggle().click();
-            expect(await timepicker.popup.isOpened()).toBeFalsy();
+            await timepicker.toggle();
+            expect(await timepicker.overlay.isOpened()).toBeFalsy();
             expect(await checkbox.isDisabled()).toBeTruthy();
             expect(await dateTimepicker.isDisabled()).toBeTruthy();
         });