+++ packages/bits/src/lib/time-frame-picker/time-frame-picker.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 
 import { NuiCommonModule } from "../../common/common.module";
-import { NuiDateTimerPickerModule } from "../date-time-picker/date-time-picker.module";
+import { NuiDateTimePickerModule } from "../date-time-picker/date-time-picker.module";
 import { NuiMenuModule } from "../menu/menu.module";
 
 import { QuickPickerComponent } from "./quick-picker/quick-picker.component";
@@ -16,7 +16,7 @@ import { TimeFramePickerComponent } from "./time-frame-picker.component";
     imports: [
         NuiCommonModule,
         NuiMenuModule,
-        NuiDateTimerPickerModule,
+        NuiDateTimePickerModule,
     ],
     declarations: [
         QuickPickerComponent,