+++ packages/charts/src/chart-donut-content/chart-donut-content.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnDestroy, OnInit } from "@angular/core";
+import { Component, Input, OnChanges, OnDestroy, SimpleChanges } from "@angular/core";
 import { Subscription } from "rxjs";
 
 import { ChartDonutContentPlugin } from "../core/plugins/chart-donut-content-plugin";
@@ -9,7 +9,7 @@ import { IElementPosition } from "../core/plugins/types";
     templateUrl: "./chart-donut-content.component.html",
     styleUrls: ["./chart-donut-content.component.less"],
 })
-export class ChartDonutContentComponent implements OnDestroy, OnInit {
+export class ChartDonutContentComponent implements OnDestroy, OnChanges {
 
     /** The plugin instance */
     @Input() public plugin: ChartDonutContentPlugin;
@@ -19,11 +19,16 @@ export class ChartDonutContentComponent implements OnDestroy, OnInit {
 
     private contentPositionUpdateSubscription: Subscription;
 
-    public ngOnInit() {
-        this.contentPositionUpdateSubscription = this.plugin.contentPositionUpdateSubject.subscribe((contentPosition: IElementPosition) => {
-            this.contentPosition = contentPosition;
-        });
-        this.plugin.chart.updateDimensions();
+    public ngOnChanges(changes: SimpleChanges) {
+        if (changes.plugin) {
+            this.contentPositionUpdateSubscription?.unsubscribe();
+
+            this.contentPositionUpdateSubscription = this.plugin.contentPositionUpdateSubject.subscribe((contentPosition: IElementPosition) => {
+                this.contentPosition = contentPosition;
+            });
+
+            this.plugin.chart.updateDimensions();
+        }
     }
 
     public ngOnDestroy() {